Move registerTools to wrapper
