+++ front/lib/actions/mcp_internal_actions/tool_definition.ts
@@ -1,4 +1,3 @@
-import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 import type { RequestHandlerExtra } from "@modelcontextprotocol/sdk/shared/protocol.js";
 import type {
   CallToolResult,
@@ -10,7 +9,6 @@ import type { z } from "zod";
 
 import type { MCPToolStakeLevelType } from "@app/lib/actions/constants";
 import type { MCPError } from "@app/lib/actions/mcp_errors";
-import { withToolLogging } from "@app/lib/actions/mcp_internal_actions/wrappers";
 import type { AgentLoopContextType } from "@app/lib/actions/types";
 import type {
   InternalMCPServerDefinitionType,
@@ -60,26 +58,6 @@ export function defineTool<TName extends string, TSchema extends ZodRawShape>(
   def: ToolDefinition<TName, TSchema>
 ): ToolDefinition<TName, TSchema> {
   return def;
-}
-
-export function registerTool(
-  auth: Authenticator,
-  server: McpServer,
-  agentLoopContext: AgentLoopContextType | undefined,
-  monitoringName: string,
-  tool: ToolDefinition
-): void {
-  server.tool(
-    tool.name,
-    tool.description,
-    tool.schema,
-    withToolLogging(
-      auth,
-      { toolNameForMonitoring: monitoringName, agentLoopContext },
-      (params, extra) =>
-        tool.handler(params, { ...extra, agentLoopContext, auth })
-    )
-  );
 }
 
 export interface ServerMetadata {
