+++ front/lib/api/actions/servers/github/index.ts
@@ -1,7 +1,7 @@
 import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 
-import { registerTool } from "@app/lib/actions/mcp_internal_actions/tool_definition";
 import { makeInternalMCPServer } from "@app/lib/actions/mcp_internal_actions/utils";
+import { registerTool } from "@app/lib/actions/mcp_internal_actions/wrappers";
 import type { AgentLoopContextType } from "@app/lib/actions/types";
 import { GITHUB_TOOL_NAME } from "@app/lib/api/actions/servers/github/metadata";
 import { createGithubTools } from "@app/lib/api/actions/servers/github/tools";
