+++ front/lib/api/actions/servers/google_calendar/index.ts
@@ -1,7 +1,7 @@
 import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
 
-import { registerTool } from "@app/lib/actions/mcp_internal_actions/tool_definition";
 import { makeInternalMCPServer } from "@app/lib/actions/mcp_internal_actions/utils";
+import { registerTool } from "@app/lib/actions/mcp_internal_actions/wrappers";
 import type { AgentLoopContextType } from "@app/lib/actions/types";
 import { GOOGLE_CALENDAR_TOOL_NAME } from "@app/lib/api/actions/servers/google_calendar/metadata";
 import { TOOLS } from "@app/lib/api/actions/servers/google_calendar/tools";
